+++ Microsoft.Azure.Cosmos/tests/Microsoft.Azure.Cosmos.EmulatorTests/CosmosContainerTests.cs
@@ -10,10 +10,8 @@ namespace Microsoft.Azure.Cosmos.SDK.EmulatorTests
     using System.IO;
     using System.Linq;
     using System.Net;
-    using System.Runtime.CompilerServices;
     using System.Threading.Tasks;
     using Microsoft.Azure.Cosmos.Resource.CosmosExceptions;
-    using Microsoft.Azure.Cosmos.Tracing;
     using Microsoft.VisualStudio.TestTools.UnitTesting;
     using Newtonsoft.Json;
     using Newtonsoft.Json.Linq;
@@ -23,18 +21,6 @@ namespace Microsoft.Azure.Cosmos.SDK.EmulatorTests
     {
         private CosmosClient cosmosClient = null;
         private Cosmos.Database cosmosDatabase = null;
-
-        [ClassInitialize]
-        public static void ClassInit(TestContext context)
-        {
-            OpenTelemetryTests.ClassInitialize(context);
-        }
-
-        [ClassCleanup]
-        public static void Cleanup()
-        {
-            OpenTelemetryTests.FinalCleanup();
-        }
 
         private static long ToEpoch(DateTime dateTime)
         {
