+++ Microsoft.Azure.Cosmos/tests/Microsoft.Azure.Cosmos.Tests/CosmosBadReplicaTests.cs
@@ -23,6 +23,18 @@ namespace Microsoft.Azure.Cosmos.Tests
     [TestClass]
     public class CosmosBadReplicaTests
     {
+        [ClassInitialize]
+        public static void ClassInit(TestContext context)
+        {
+            OpenTelemetryTests.ClassInitialize(context);
+        }
+
+        [ClassCleanup]
+        public static void FinalCleanup()
+        {
+            OpenTelemetryTests.FinalCleanup();
+        }
+
         [TestMethod]
         [Timeout(30000)]
         public async Task TestGoneFromServiceScenarioAsync()
