+++ Microsoft.Azure.Cosmos/tests/Microsoft.Azure.Cosmos.Tests/CosmosClientTests.cs
@@ -32,6 +32,18 @@ namespace Microsoft.Azure.Cosmos.Tests
     {
         public const string AccountEndpoint = "https://localhost:8081/";
         public const string ConnectionString = "AccountEndpoint=https://localtestcosmos.documents.azure.com:443/;AccountKey=425Mcv8CXQqzRNCgFNjIhT424GK99CKJvASowTnq15Vt8LeahXTcN5wt3342vQ==;";
+        
+        [ClassInitialize]
+        public static void ClassInit(TestContext context)
+        {
+            OpenTelemetryTests.ClassInitialize(context);
+        }
+
+        [ClassCleanup]
+        public static void FinalCleanup()
+        {
+            OpenTelemetryTests.FinalCleanup();
+        }
 
         [TestMethod]
         public async Task TestDispose()
